Crate pallet_bags_list_remote_tests

Utilities for remote-testing pallet-bags-list.

Modules

migration

Test to check the migration of the voter bag.

sanity_check

Test to execute the sanity-check of the voter bag.

snapshot

Test to execute the snapshot using the voter bag.

Constants

LOG_TARGET

A common log target to use.

Traits

RuntimeT

A wrapper for a runtime that the functions of this crate expect.

Functions

display_and_check_bags

Display the number of nodes in each bag, while identifying those that need a rebag.